Improve responsiveness of layout
This commit converts many absolute units to relative ones to make the
layout respond better to changes in screen size. The placement of the
elements other than the board now depends on the height of the screen.
If the viewport’s height is relatively small, the board takes up all the
vertical space and the help area is hidden. If the viewport is wider
than it is high, the information area is placed at the side of the
board.